+++ packages/suite/src/views/wallet/account/receive/components/AddressItem/index.tsx
