+++ packages/suite/src/views/wallet/account/receive/components/QrCode/messages.ts
